code: ajustados erros.

+++ src/main/java/biblioteca_de_jogos/classes/Reserva.java
@@ -37,7 +37,7 @@ public class Reserva {
 		return this.idUsuario;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 }
